Delete comment added

+++ public/js/dashboard.js
@@ -5,23 +5,29 @@ const newPostButton = async (event) => {
 
 const deletePostButton = async (event) => {
   event.preventDefault();
-  const post_id = event.target.id;
 
-  const response = await fetch(`/api/posts/${post_id}`, {
-    method: 'DELETE',
-  });
-  if(response.ok) {
-    document.location.reload();
-  } else {
-    alert(response.statusText);
+  const result = confirm('Are you sure you want to delete this post?')
+
+  if (result) {
+
+    const post_id = event.target.id;
+
+    const response = await fetch(`/api/posts/${post_id}`, {
+      method: 'DELETE',
+    });
+    if (response.ok) {
+      document.location.reload();
+    } else {
+      alert(response.statusText);
+    }
   }
+
 }
 
 const editPostButton = async (event) => {
   event.preventDefault();
 
   const post_id = event.target.id;
-  console.log(event.target);
   
   document.location.assign(`editpost/${post_id}`)
 }
